requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
email-validator==2.1.0
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
